Refaktorere barn-skjema
Gruppere barn-felter i egen komponent
